# Conflicts:
#	src/main/java/com/spring/gradle/defect/controller/DefectController.java
#	src/main/java/com/spring/gradle/defect/controller/ModuleController.java
#	src/main/java/com/spring/gradle/defect/controller/ProjectController.java
#	src/main/java/com/spring/gradle/defect/controller/ReleaseController.java
#	src/main/java/com/spring/gradle/defect/controller/SmokeTestController.java
#	src/main/java/com/spring/gradle/defect/controller/StatusController.java
#	src/main/java/com/spring/gradle/defect/controller/TeamMemberController.java
#	src/main/java/com/spring/gradle/defect/controller/UserController.java
#	src/main/java/com/spring/gradle/defect/dto/ProjectDto.java
#	src/main/java/com/spring/gradle/defect/dto/StatusDto.java
#	src/main/java/com/spring/gradle/defect/entity/Module.java
#	src/main/java/com/spring/gradle/defect/entity/Project.java
#	src/main/java/com/spring/gradle/defect/entity/SmokeTest.java
#	src/main/java/com/spring/gradle/defect/entity/Status.java
#	src/main/java/com/spring/gradle/defect/entity/User.java
#	src/main/java/com/spring/gradle/defect/repository/UserRepository.java
#	src/main/java/com/spring/gradle/defect/service/DefectService.java
#	src/main/java/com/spring/gradle/defect/service/SmokeTestService.java
#	src/main/java/com/spring/gradle/defect/service/StatusService.java
#	src/main/java/com/spring/gradle/defect/service/implementation/DefectImplementation.java
#	src/main/java/com/spring/gradle/defect/service/implementation/ModuleServiceImplementation.java
#	src/main/java/com/spring/gradle/defect/service/implementation/ProjectServiceImplementation.java
#	src/main/java/com/spring/gradle/defect/service/implementation/StatusServiceImplementation.java
#	src/main/java/com/spring/gradle/defect/service/implementation/UserImplementation.java

+++ src/main/java/com/spring/gradle/defect/repository/StatusRepository.java
@@ -3,6 +3,9 @@ package com.spring.gradle.defect.repository;
 import com.spring.gradle.defect.entity.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StatusRepository extends JpaRepository<Status, Long> {
+public interface StatusRepository extends JpaRepository<Status, Integer> {
 
+	//void save(StatusDto statusDto);
+
+	
 }
